requirements.txt
fastapi
motor
uvicorn
python-dotenv
